Migrate cost_usd from float to Decimal
Real per-record costs go below a cent (a live Anthropic pull came back
$0.00501), and float summation drifts by summation order — harmless for a
read-only v0.1 report but not once a future pacing/enforcement feature
makes spend decisions on this number. schema.py's cost_usd TODO(pre-v0.2)
called this out explicitly.

cost_usd itself, pricing.yaml rates, and every dollar-amount field derived
from them (daily totals, forecast, overspend projection, batch/tier gap,
reconciliation totals) are now Decimal end to end, including the JSON
cache round-trip (written as a string — json has no Decimal type).
Ratios that aren't dollar amounts (breakdown share, pct_change, z-score,
output_token_share) stay float, converted right after the division that
produces them.

Verified against real cached data (terminal, HTML, --share) and both
synthetic scales (personal spike z=35.1, team spike z=9.4 at the intended
seed) — no regressions in anomaly detection, batch gap, or reconciliation.

+++ src/llm_spend/analysis/forecast.py
@@ -12,6 +12,7 @@ from __future__ import annotations
 import calendar
 from dataclasses import dataclass
 from datetime import date
+from decimal import Decimal
 
 from llm_spend.analysis._timeseries import daily_totals
 from llm_spend.schema import UsageRecord
@@ -27,9 +28,9 @@ DISCLAIMER = (
 class ForecastResult:
     days_elapsed: int
     days_in_month: int
-    spend_so_far: float
-    daily_average: float
-    projected_total: float
+    spend_so_far: Decimal
+    daily_average: Decimal
+    projected_total: Decimal
     disclaimer: str = DISCLAIMER
